requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
bitarray==2.9.2
certifi==2023.11.17
charset-normalizer==3.3.2
cytoolz==0.12.2
eth-account==0.10.0
eth-hash==0.6.0
eth-keyfile==0.7.0
eth-keys==0.5.0
eth-rlp==1.0.0
eth-typing==4.0.0
eth-utils==3.0.0
eth_abi==5.0.0
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.6
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
multidict==6.0.4
parsimonious==0.9.0
protobuf==4.25.2
pycryptodome==3.20.0
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.0
pyunormalize==15.1.0
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rlp==4.0.0
rpds-py==0.17.1
toolz==0.12.0
typing_extensions==4.9.0
urllib3==2.1.0
web3==6.14.0
websockets==12.0
yarl==1.9.4
